Created Log In Page and added Back buttons
Created the Log In page and added back buttons to both the login and sign up pages.

I also added a ScrollView to all 3 pages we have so far. I did this mainly so that your keyboard disappears when you click out of a textbox. So I guess we don't need it on the title screen but I remember someone said their "Sign Up" button was too far down on their screen and they couldn't press it. So hopefully they can at least move the screen now and press it.