Got rid of Warning; Changed Collaborate to use Popover instead of Modal to display the Create a New Team dialog
This was the warning "Warning: Cannot update during an existing state transition (such as within render). Render methods should be a pure function of props and state."

For some reason using Modal didn't work. It seemed to break (throw a warning) on trying to change the Input and set the tempTeamName. Not sure why because changing it to use Popover instead worked fine.